Initial rename from regulation to control_list
